[FAB-1362] Add KafkaBrokers to shared config
https://jira.hyperledger.org/browse/FAB-1362

The list of Kafka brokers used for ordering needs to be shared across the
shims (ordering service nodes). This changeset introduces the key and
the getter of that key's value to the sharedconfig package. It also adds
the necessary definition in the orderer protos.

This value is not yet encoded as a configuration item in the genesis
block that the static bootstrapper produces. This will come at a
subsequent changeset, with the introduction of the provisional
bootstrapper.

Change-Id: I6268e79dc25ca2cee8da248ba44c9a572c04d909
Signed-off-by: Kostas Christidis <kostas@christidis.io>