FAB-1018 MultiChannel API fabric<-->gossip
This commit changes the API the fabric provides to the gossip layer
to conform to latest discussions.
Previously the JoinChannel message was expected to have all peers in the channel,
 now it can have only the anchor peers.
In addition, each peer needs to know in which organization
a fellow peer is, in order to know whether its safe to pass
a block to it or not via state transfer.

Change-Id: I6af5ce21a14bf6c4eaf1416aefe2877a147f15ec
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>