[FAB-1352] Add time-based block cutting to Kafka
https://jira.hyperledger.org/browse/FAB-1352

In the version that was rebased on top of the common components, this
option was kept out in order to minimize the complexity of the changeset.

This changeset introduces it so now the Kafka-based consenter uses the
BatchTimeout setting in shared config, and posts time-to-cut (TTC-X)
messages according to the design document posted here:
https://docs.google.com/document/d/1vNMaM7XhOlu9tB_10dKnlrhy5d7b1u8lSY8a-kVjCO4/edit

The respective unit tests from the solo package have been ported, as well
as additional tests specific to the time-to-cut logic. Do note that this
path shall be revisited with integration tests.

Change-Id: I743d4412cf8a3536fcb854433dfcbb3baa221d95
Signed-off-by: Kostas Christidis <kostas@christidis.io>