[FAB-1449] Add getPkiID to the commReceivedMsg
Some messages in gossip aren't signed,
because they are sent only point-to-point
so the identity of the sender can be extracted
from the PKI-ID that's sent during the comm layer authentication.

In order to support multi-channel related access control
for messages that are point-to-point I need to add getPkiID to the interface.
The idea is that the gossip layer would drop point-to-point
messages that came from peers that are not in the channel.

Change-Id: I0468a26764366fb5d02f8db846969209eea024d3
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>