fix sbft consensus violation after attack
A correct replica would blindly put together a checkpointed batch with
its locally preprepared one for the same sequence number. However, even
correct replicas need to check if the payloads are the same, as Byzantine
primary might had previously polluted a correct replica with a different
preprepare.

Added a test that was causing consensus violation in the previous code
(now passing).

Change-Id: I5417730cc95d05c619f445178fa93ecfe5043a70
Signed-off-by: Marko Vukolic <mvu@zurich.ibm.com>