[FAB-1333] Make orderer logging configurable
https://jira.hyperledger.org/browse/FAB-1333

The orderer code has taken an ad-hoc approach to logging during
construction.  This changeset unifies the logging across the common
components, solo, and kafka.  It deliberately does not touch the sbft
portion of the package so as not to disrupt its development until it is
ready to move onto the common config and common code.

Note that this moves flogging into the common package, and splits the
LoggingInit into two pieces.  The peer assumes a single global viper and
peer specific config structures.  This dependency leaked into flogging,
but in reality, the spec should probably be passed from the peer to
flogging, rather than having flogging interrogate the configuration
itself.  This would however have balooned the changeset, so it is left
as is for now.

Revision 3: Per feedback from Kostas, fixed the stuttering which was
prevalent in the flogging package.
Revision 4: Fix linting errors
Revision 6: Fix rebasing bugs which balooned the diff

Change-Id: I3ed21a3a001fb1e9c6b28be700203b84c69ee9f0
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>