[FAB-1573] Make deliver API signable
https://jira.hyperledger.org/browse/FAB-1573

This changeset modifies the Deliver API to accept an Envelope of type
DELIVER_SEEK_INFO instead of a standard SeekInfo message.  The Envelope
embeds a SeekInfo as well as the bits necessary to construct a
non-replayable signature.

This unfortunately breaks the API, but it seems better to fix it now
than to wait until there are more dependencies.

The alternative was to invent yet another signing scheme, which seemed
like it would be equally as much work on the consumer side with ongoing
messiness.

Change-Id: I3e8a759d33f68cb6a69f5242440136da730f0396
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>