Fix go_package in attributes.proto
At some point, the attributes.proto file appears to have moved, causing
the go_package directive to point to the wrong location.  This causes
the attributes.pb.go file to be generated at the old location, and does
not actually update the utilized copy.

This changeset fixes the go_package and rebuilds the proto.

Change-Id: I1b2df6a0056ca458decfa56b07eb56255f4bf67c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>