Fix building COP in vagrant
See https://jira.hyperledger.org/browse/FAB-1597 for this issue.
The problem is that we could not build COP in vagrant.
The reason was that one of COP's vendored packages has native code
with an 'import "C"' statement in it which uses cgo, which uses the
CGO_LDFLAGS env variable to link, which had "-lrocksdb" on it.

This change set removes CGO_LDFLAGS from devenv/setup.sh

While this change set fixes the COP build issue, there are other
references to rocksdb in fabric examples which need cleaning up.
See https://jira.hyperledger.org/browse/FAB-1636

Change-Id: I54e877b2c1f872811a18e983b1e1c05eb5ab1700
Signed-off-by: Keith Smith <bksmith@us.ibm.com>