Orderer Signer MSP-based implementation
This change-set introduces an MSP-based
orderer multichain.signer implementation.

At bootstrap, an orderer loads his local
signing identity that is then used to sign
messages by means of the multichain.Signer interface.

This change-set comes in the context of
https://jira.hyperledger.org/browse/FAB-1268

Also updates the orderer config path stuff to be slightly more
consistent to support referencing the msp sampleconfig both from the dev
env and the docker image.

Change-Id: Ie4e04cb18d0f28ce801a474a3c500d4f61fa9c7f
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>