[FAB-1765] Fix orderer crash
https://jira.hyperledger.org/browse/FAB-1765

When the orderer is initialized with an empty slice for
ChainCreationPolicyNames, no chain is initialized as the system chain.
This does not cause a panic, and leads to a nil pointer dereference when
a new chain proposal tries to reference the system chain.

This is a byproduct of the fact that empty slices marshal to nil under
protobuf, so the sharedconfig code was setting the
ChainCreationPolicyNames to nil rather than an empty slice. The system
chain was detected by looking for a non-nil ChainCreationPolicyNames.

Change-Id: I4c8bf40e89dc73ba4ee27646b3f66c65ecb22d38
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>