FAB-1585 Refactor History to enable LevelDB
The purpose of this refactor is to create a common interface and to
isolate the db implementation, so that we can enable history on
LevelDB. This change only refactors the existing code and tests.
It does not add new functionality or any new tests. There will be
future changes that add history support on LevelDB and clean up
tests to work against LevelDB.

Note: History is not enabled by default. History must be enabled in
core.yaml historyDatabase for it's unit tests to run.

Change-Id: Ia5293cd345be4499e5e3df0c2949af16f71a608d
Signed-off-by: denyeart <enyeart@us.ibm.com>