[FAB-1652] Switch runtime to baseos
The peer/orderer containers are currently built on a custom
busybox base we call "runtime".  This results in superbly small
images (~25MB each).  However, it was discovered that we also
inadvertently broke DNS/gethostbyname() functions in glibc
as a result of doing this.  It was decided that the busybox
path, while small, was not a tenable solution for the long
term.  Therefore, we split the baseimage up into three
pieces: baseos, basejvm, and (the original) baseimage.

BaseOS gives us a relatively small os (64MB compressed for
release x86_64-0.3.0) but without the problematic DNS/glibc
issues.

Therefore, this patch removes our custom busybox build
from fabric.git and instead builds the peer/orderer
from hyperledger/fabric-baseos, which is now available
in dockerhub.  This has the side benefit of also speeding
up the build since we no longer need to compile busybox.

Change-Id: I0dd1c61ea5afd10402cf6830f8fc605378a259a7
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>