[FAB-1885] GetTransactionByID to return Tran Envelope
QSCC GetTransactionByID should return a Transaction  Envelope,
not a Transaction. The Envelope contains the signature and
Payload, which contains the transaction header and Transaction.

Clients will want the transaction header information when retrieving
the transaction.  They may also want the signature to verify that
the transaction hasn't changed since the time they submitted it.
The return type will change from Transaction proto to Envelope proto.

Change-Id: I1de238035c3b6cd00abb3ed8506c54566ee0f2b0
Signed-off-by: denyeart <enyeart@us.ibm.com>