[FAB-678] Add BuildSpecFactory to container controller
We currently take an io.Reader as a parameter to
Controller.Start().  This is relatively inconsequential since the
docker tarball is precomputed.  However, in the near future, we
will be splitting this up so that cds.CodePackage is only a src
container.  The formulation into a docker container will happen
JIT in response to a peer trying to start the container.  Once
this happens, we want to avoid the overhead of computing the
build specification unless it is actually needed.  Therefore,
we use a callback that the docker-controller may use to indicate
when the build-spec is needed.

Part of FAB-678

Change-Id: I259c9c0b600866c2b78c4319039ca8214d2a9620
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>