[FAB-1349] Enforce restrictions on chain IDs
https://jira.hyperledger.org/browse/FAB-1349

This changeset introduces a function which rejects chain IDs (i.e.
channel names) that do NOT comply with the following restrictions:

1. Contain only ASCII alphanumerics, dots '.', dashes '-'.
2. Are shorter than 250 characters.
3. Are not "." or "..".

Our hand here is forced by Kafka, since a channel in Fabric maps to a
Kafka topic, and there are restrictions for the allowed topic names [1].

Note that underscores are allowed in Kafka, but topics with a period or
underscore could collide. (Thanks to Luis Sanchez for bringing this to my
attention.) We therefore remove support for underscores altoghether to
keep the checks simpler.

UPDATE: Switched from a filter to a function because filters are not
evaluated during the system chain creation.

[1] https://github.com/apache/kafka/blob/trunk/core/src/main/scala/kafka/common/Topic.scala#L29

Change-Id: I14b95477e485fea27868338e2f33772588b8a770
Signed-off-by: Kostas Christidis <kostas@christidis.io>