Store hashes in rwset for phantom reads issue
https://jira.hyperledger.org/browse/FAB-1935

This CR enables storing of hashes instead of the full read-set for a
range query as the range query results can be significant large in
some cases.

More detail is captured in the Jira item.

I tried to contain this CR to some limit but seems unavoidable because,
in addition to new functionality, it requires a number of tests and
detailed comments. I have tried to put the new code in new files and
minimize the changes to the existing files so as to make the review
relatively easy.

Change-Id: I1c89272be73cde397279d965e4c78020ea2cd696
Signed-off-by: manish <manish.sethi@gmail.com>