Get ID for local MSP from core.yaml
This change-set removes the hardcoded ID of the local MSP and reads it from
core.yaml instead. Additionally, the code that sets up the MSPs for the test
chain has been cleaned up - the test setup is now only used in testing and
the peer always sets up chain MSPs from the config block (whether it's the
test chain or a chain it is asked to join).

Change-Id: Ife55c9be831172e603b4495bdfa37982a571d3a9
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>